Edit view livewire crm customer-type-list

+++ resources/views/livewire/crm/customer-type/customer-type-lists.blade.php
@@ -23,19 +23,19 @@
                     </div>
                     <div class="col-3 d-flex justify-content-center">
                         <div class="btn-group w-100" role="group">
-                            <button wire:click="$dispatch('add-customer-type')" type="button" class="btn btn-primary"
-                                data-toggle="modal" data-target="#modal-add-customer-type">
-                                <i class="fas fa-plus"></i>
-                            </button>
 
-                            <!-- Customer Type Add Component -->
-                            @livewire('crm.customer-type.customer-type-create')
-                            <!-- ./Customer Type Add Component -->
+                            @can('customerType.create')
+                                <button wire:click="$dispatch('add-customer-type')" type="button" class="btn btn-primary"
+                                    data-toggle="modal" data-target="#modal-add-customer-type">
+                                    <i class="fas fa-plus"></i>
+                                </button>
+                            @endcan
 
                             <button wire:click="$dispatch('refresh-customer-type')" type="button"
                                 class="btn btn-success">
                                 <i class="fas fa-sync-alt"></i>
                             </button>
+
                             <button type="button" class="btn btn-warning" data-dismiss="modal">
                                 <i class="fas fa-times-circle"></i>
                             </button>
@@ -56,14 +56,18 @@
                 </div>
 
                 <div class="table-responsive">
-                    <table class="table table-hover table-sm">
-                        <thead>
+                    <table class="table table-sm table-hover table-bordered">
+                        <thead class="table-info">
                             <th scope="col">#</th>
                             <th scope="col">Created</th>
                             <th scope="col">Updated</th>
                             <th scope="col">Id</th>
                             <th scope="col">Customer Type Name</th>
-                            <th scope="col" style="width: 90px">Action</th>
+                            @can('customerType.edit')
+                                <th scope="col" colspan="2" class="text-center">Action</th>
+                            @elsecan('customerType.delete')
+                                <th scope="col" colspan="2" class="text-center">Action</th>
+                            @endcan
                         </thead>
 
                         <tbody>
@@ -107,28 +111,37 @@
                                     </td>
                                     <td>{{ $item->id }}</td>
                                     <td>{{ $item->name }}</td>
-                                    <td>
-                                        <button
-                                            wire:click.prevent="$dispatch('edit-customer-type',{id:{{ $item->id }}})"
-                                            type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
-                                            data-target="#modal-edit-customer-type">
-                                            <i class="fas fa-edit"></i>
-                                        </button>
-                                        <button
-                                            wire:click.prevent="deleteCustomerType({{ $item->id }},{{ "'" . str_replace("'", '', $item->name) . "'" }})"
-                                            class="btn btn-sm btn-danger">
-                                            <i class="fas fa-trash"></i>
-                                        </button>
-                                    </td>
+
+                                    @can('customerType.edit')
+                                        <td style="width: 40px" class="p-1 text-center">
+                                            <button
+                                                wire:click.prevent="$dispatch('edit-customer-type',{id:{{ $item->id }}})"
+                                                type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
+                                                data-target="#modal-edit-customer-type">
+                                                <i class="fas fa-edit"></i>
+                                            </button>
+                                        </td>
+                                    @endcan
+
+                                    @can('customerType.delete')
+                                        <td style="width: 40px" class="p-1 text-center">
+                                            <button
+                                                wire:click.prevent="deleteCustomerType({{ $item->id }},{{ "'" . str_replace("'", '', $item->name) . "'" }})"
+                                                class="btn btn-sm btn-danger">
+                                                <i class="fas fa-trash"></i>
+                                            </button>
+                                        </td>
+                                    @endcan
                                 </tr>
                             @endforeach
 
-                            <!-- Customer Type Edit Component -->
                             @livewire('crm.customer-type.customer-type-edit')
-                            <!-- ./Customer Type Edit Component -->
 
                         </tbody>
                     </table>
+
+                    @livewire('crm.customer-type.customer-type-create')
+
                 </div>
 
             </div>
@@ -148,9 +161,6 @@
 @script
     <script>
         $wire.on("confirm-delete-customer-type", (event) => {
-
-            // alert(event.name);
-
             Swal.fire({
                 title: "Are you sure delete ?",
                 text: `Customer Type : ${event.name}`,
@@ -174,7 +184,6 @@
     <script>
         document.addEventListener('livewire:initialized', () => {
             @this.on('close-modal-customer-type', (event) => {
-                // alert('Close Modal')
                 setTimeout(() => {
                     $('#modal-edit-customer-type').modal('hide')
                 }, 3000);
